requirements.txt
nsetools 
nselib
pandas
numpy
matplotlib
yfinance
feedparser
vaderSentiment
pymupdf
langchain
chromadb
openai
tiktoken
langchain-openai 
langchain-chroma
lime
torch
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
certifi
pip-system-certs
idna
urllib3==2.6.2
requests
faiss-cpu
scikit-learn